+++ CHANGELOG.md
@@ -5,13 +5,21 @@ of [keepachangelog.com](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [0.3.1] - 2021-10-22
+
+### Added
+
+- Added `str->local-date` function were you can use to convert `String` to `LocalDate`.
+
+- Added `local-date->str` function were you can use to convert `LocalDate` to `String`.
+
 ## [0.2.1] - 2021-09-06
 
 ### Added
 
 - Applying `loose-schema` function to a schema with inner schemas, also make theirs inner schemas loose.
 
-## 0.1.1 - 2021-09-05
+## [0.1.1] - 2021-09-05
 
 ### Fixed
 
@@ -24,6 +32,8 @@ of [keepachangelog.com](http://keepachangelog.com/).
 - Add `loose-schema` function.
 
 [Unreleased]: https://github.com/macielti/common-clj/compare/0.1.1...HEAD
+
+[0.3.1]: https://github.com/macielti/common-clj/compare/0.3.1...0.2.1
 
 [0.2.1]: https://github.com/macielti/common-clj/compare/0.2.1...0.1.1